refactor: 공통 리소스를 shared 폴더로 이동 및 구조 정리
