+++ src/features/feed/api/feedApi.js
@@ -1,11 +1,11 @@
 // src/api/feed.js
-import axios from "@/api/axiosInstance";
+import { axiosInstance } from "@/shared";
 import { parseFeedsResponse, parseFeedDetailResponse } from "@/features/feed";
 
 // ✅ feed 전체 조회
 export const getFeeds = async (page = 1, size = 10) => {
   try {
-    const res = await axios.get(`/api/feed`, {
+    const res = await axiosInstance.get(`/api/feed`, {
       params: { page, size },
     });
     if (res.status === 200 && res.data?.data) {
@@ -65,7 +65,7 @@ export const getFeeds = async (page = 1, size = 10) => {
 // ✅ feed 상세 조회
 export const getFeedDetail = async (feedId) => {
   try {
-    const res = await axios.get(`/api/feed/${feedId}`);
+    const res = await axiosInstance.get(`/api/feed/${feedId}`);
     if (res.status === 200 && res.data?.data) {
       return parseFeedDetailResponse(res.data.data);
     }
@@ -120,37 +120,40 @@ export const getFeedDetail = async (feedId) => {
 
 // ✅ feed 등록
 export const createFeed = async (data) => {
-  const response = await axios.post("/api/feed", data);
+  const response = await axiosInstance.post("/api/feed", data);
   return response.data;
 };
 
 // ✅ feed 수정
 export const updateFeed = async (feedId, data) => {
-  const response = await axios.patch(`/api/feed/${feedId}`, data);
+  const response = await axiosInstance.patch(`/api/feed/${feedId}`, data);
   return response.data;
 };
 
 // ✅ feed 삭제
 export const deleteFeed = async (feedId) => {
-  const response = await axios.delete(`/api/feed/${feedId}`);
+  const response = await axiosInstance.delete(`/api/feed/${feedId}`);
   return response.data;
 };
 
 // ✅ 댓글 조회
 export const getComments = async (feedId) => {
-  const response = await axios.get(`/api/feed/${feedId}/comments`);
+  const response = await axiosInstance.get(`/api/feed/${feedId}/comments`);
   return response.data;
 };
 
 // ✅ 댓글 등록
 export const createComment = async (feedId, data) => {
-  const response = await axios.post(`/api/feed/${feedId}/comments`, data);
+  const response = await axiosInstance.post(
+    `/api/feed/${feedId}/comments`,
+    data
+  );
   return response.data;
 };
 
 // ✅ 댓글 수정
 export const updateComment = async (feedId, commentId, data) => {
-  const response = await axios.patch(
+  const response = await axiosInstance.patch(
     `/api/feed/${feedId}/comments/${commentId}`,
     data
   );
@@ -159,7 +162,7 @@ export const updateComment = async (feedId, commentId, data) => {
 
 // ✅ 댓글 삭제
 export const deleteComment = async (feedId, commentId) => {
-  const response = await axios.delete(
+  const response = await axiosInstance.delete(
     `/api/feed/${feedId}/comments/${commentId}`
   );
   return response.data;
