+++ src/features/feed/modals/FeedPostModal.jsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import PostForm from "@/components/modals/PostForm";
-import useLockBodyScroll from "@/hooks/useLockBodyScroll";
+import { useLockBodyScroll } from "@/shared";
 import { createFeed } from "@/features/feed";
 
 const FeedPostModal = ({ onClose, onPostCreate }) => {
