+++ src/shared/index.js
@@ -1,0 +1,18 @@
+// ✅ API
+export { default as axiosInstance } from "./api/axiosInstance";
+
+// ✅ Components
+export { default as Header } from "./components/layout/Header";
+export { default as Sidebar } from "./components/layout/Sidebar";
+export { default as Filterbar } from "./components/Filterbar";
+export { default as FallbackImage } from "./components/FallbackImage";
+export { default as ProfileImage } from "./components/ProfileImage";
+
+// ✅ Hooks
+export { default as useLockBodyScroll } from "./hooks/useLockBodyScroll";
+
+// ✅ Utils
+export * from "./utils/regionMapper";
+
+// ✅ Constants
+export { default as regionOptions } from "./constants/regionOptions";
